Extract holographicMaterial instance
