+++ app/_materials/holographicMaterial.ts
@@ -1,0 +1,4 @@
+import { Material } from "three";
+import { HolographicMaterial } from "./HolographicMaterialConstructor";
+
+export const holographicMaterial = new HolographicMaterial() as Material;
